requirements.txt
streamlit
pandas
numpy
matplotlib
mplsoccer
Pillow
requests
seaborn
scipy
adjustText
supabase
unidecode